+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+[v0.5]
+------
+
+### Added
+- `scopeIdProvider` property on RateLimiterComponent allows you to override the
+  default scope ID used for rate allowance storage.
+
+
 [v0.4]
 ------